+++ relay-config/src/config.rs
@@ -1338,10 +1338,6 @@ pub enum EmitOutcomes {
     AsClientReports,
     /// Emit outcomes as outcomes.
     AsOutcomes,
-    /// Emit outcomes as metrics.
-    ///
-    /// This is a temporary and experimental option which should not be used and will be removed again.
-    AsMetrics,
 }
 
 impl EmitOutcomes {
@@ -1361,7 +1357,6 @@ impl Serialize for EmitOutcomes {
             Self::None => serializer.serialize_bool(false),
             Self::AsClientReports => serializer.serialize_str("as_client_reports"),
             Self::AsOutcomes => serializer.serialize_bool(true),
-            Self::AsMetrics => serializer.serialize_str("as_metrics"),
         }
     }
 }
@@ -1372,7 +1367,7 @@ impl Visitor<'_> for EmitOutcomesVisitor {
     type Value = EmitOutcomes;
 
     fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        formatter.write_str("true, false, 'as_client_reports' or 'as_metrics'")
+        formatter.write_str("true, false, 'as_client_reports'")
     }
 
     fn visit_bool<E>(self, v: bool) -> Result<Self::Value, E>
@@ -1392,7 +1387,6 @@ impl Visitor<'_> for EmitOutcomesVisitor {
     {
         match v {
             "as_client_reports" => Ok(EmitOutcomes::AsClientReports),
-            "as_metrics" => Ok(EmitOutcomes::AsMetrics),
             _ => Err(E::invalid_value(Unexpected::Str(v), &self)),
         }
     }