+++ tests/integration/test_spans.py
@@ -9,7 +9,7 @@ from requests import HTTPError
 from sentry_relay.consts import DataCategory
 from sentry_sdk.envelope import Envelope, Item, PayloadRef
 
-from .asserts import matches_any
+from .asserts import time_within_delta
 from .test_store import make_transaction
 
 TEST_CONFIG = {
@@ -1083,7 +1083,7 @@ def test_dynamic_sampling(
     if sample_rate == 1.0:
         spans = spans_consumer.get_spans(timeout=10, n=3)
         assert len(spans) == 3
-        outcomes = outcomes_consumer.get_outcomes(timeout=10, n=3)
+        outcomes = outcomes_consumer.get_outcomes(timeout=10, n=1)
         assert summarize_outcomes(outcomes) == {(16, 0): 3}  # SpanIndexed, Accepted
     else:
         outcomes = outcomes_consumer.get_outcomes(timeout=10, n=1)
@@ -1229,22 +1229,16 @@ def test_outcomes_for_trimmed_spans(mini_sentry, relay):
     assert outcomes == [
         {
             "category": DataCategory.SPAN,
-            "key_id": 123,
-            "org_id": 1,
             "outcome": 3,  # invalid
-            "project_id": 42,
             "quantity": 1,
             "reason": "too_large:span",
-            "timestamp": matches_any(),
+            "timestamp": time_within_delta(),
         },
         {
             "category": DataCategory.SPAN_INDEXED,
-            "key_id": 123,
-            "org_id": 1,
             "outcome": 3,  # invalid
-            "project_id": 42,
             "quantity": 1,
             "reason": "too_large:span",
-            "timestamp": matches_any(),
+            "timestamp": time_within_delta(),
         },
     ]